+++ src/app/controller/notfound.go
@@ -9,7 +9,8 @@ import (
 // NotFound is called when a non-existent resource is requested.
 func NotFound(w http.ResponseWriter, r *http.Request) {
 	res := model.Error{
-		Message: "Resource not found.",
+		Message: "No cheating! I'm watching you!",
+		Error: "resource not found.",
 	}
 
 	sendJson(w, http.StatusNotFound, res)